CLI: fix `logout`, update generated code
During a previous codegen, functionality related to `logout` was erased.
This command is meant to accept an arbitrary dashboard token and expire
it, not the token configured for the client itself for regular API
calls.

This diff also applies some pending changes made in the code generator
templates, designed with the aim of reducing friction during codegen and
as well as the number of hand-edits required to get the project building
post-codegen.

- adding optional imports (which clippy can trivially remove as needed).
- combining the "types" and "commands" into a single file.
- special casing `before` and `after` so we don't have to go in and
  convert them to strings after the codegen runs.

Some hand-edits are still required to remove things that are not exposed
in the SDK, but this should reduce some friction at least.
